Cleanup, remove line that causes HardFault
I finally undertand why the following line in "CDC_Receive_FS" causes a HardFault:

`RX_FIFO.data[RX_FIFO.head]=*Buf++;`

The issue is that the Buf pointer gets incremented everytime a data element is pushed to RX_FIFO.data. This effectively modifies the address where  data is written by the rest of the USB stack. Depending on the size of the array used to initialize USBD_CDC_SetRxBuffer, sooner or later the buffer will be exhausted and random data in RAM will be overwritten; this is what causes the HardFault.